fix(movement): improve singleton to allow uncomplete instance.
This could be weird, but if we want to fetch the currentState before loading the default game board, this is the onyl way we found. If no instance is created, and player constructor parameters is not given, then, create a simple class with only keys binding map. And if another constructor is called with player, it is automatically set to the current instance if one is defined, else the player is assigned to the first one.

Fix(Movement): Fixing infinite loop while attempting to fully randomize movement keys, now it have a limit of 10 attempts per keys